authentication part is done for now

+++ Authentication/views.py
@@ -8,12 +8,6 @@ from .serializers import CreateUserSerializer
 import random
 from django.core.mail import send_mail
 
-l = {
-"email":"admin@admin.com",
-"password": "aaadmin",
-"code": 1222
-}
-
 class CreateUser(APIView):
     permission_classes = [AllowAny]
 
@@ -28,33 +22,38 @@ class CreateUser(APIView):
 class ResetCode(APIView):
     permission_classes = [AllowAny]
 
-    def get(self, request, email):
-        user = MyUser.object.get(email=email)
-        verify = VerificationCode.objects.get(userId=user.id, codeType='Reset')
-        if user:
-            if verify: verify.delete()
-            generated_value = random.randrange(1000, 10000)
-            keepValue = VerificationCode.objects.get_or_create(codeType='Reset', code=generated_value, userId=user)
-            send_mail(
-                'Password Reset',
-                f'Hello, {user.first_name}. This is your password reset code {generated_value}.\
-                     Enter this into the field provided and set a new password',
-                'admin@admin.com',
-                [email],
-                fail_silently=False,
-            )
-            return Response(status=status.HTTP_202_ACCEPTED)
-        return Response(status=status.HTTP_404_NOT_FOUND)
-
-    def post(self, request, email=None):
-        user = MyUser.object.get(email=request.data['email'])
-        codeB = VerificationCode.objects.get(userId=user.id, codeType="reset")
-        if codeB.code == request.data['code']:
-            reg_serializer = CreateUserSerializer(user, data=request.data, partial=True)
-            if reg_serializer.is_valid(raise_exception=True):
-                newuser = reg_serializer.save()
-                codeB.delete()
-                if newuser:
-                    return Response(status=status.HTTP_201_CREATED)
-            return Response(reg_serializer.errors)
-        return  Response(data={"message":"this is not the right code"})
+    def post(self, request, code_or_reset=None):
+        if code_or_reset == "reset":
+            user = MyUser.object.get(email=request.data['email'])
+            codeB = VerificationCode.objects.get(userId=user.id, codeType="reset")
+            if codeB.code == request.data['code']:
+                reg_serializer = CreateUserSerializer(user, data=request.data, partial=True)
+                if reg_serializer.is_valid(raise_exception=True):
+                    newuser = reg_serializer.save()
+                    codeB.delete()
+                    if newuser:
+                        return Response(status=status.HTTP_201_CREATED)
+                return Response(reg_serializer.errors)
+            return  Response(status=status.HTTP_400_BAD_REQUEST, data={"message":"this is not the right code"})
+        elif code_or_reset == "code":
+            email = request.data["email"]
+            try:
+                user = MyUser.object.get(email=email)
+            except: user = ''
+            if user:
+                try:
+                    verify = VerificationCode.objects.get(userId=user.id, codeType='Reset')
+                except: verify = ""
+                if verify: verify.delete()
+                generated_value = random.randrange(1000, 10000)
+                keepValue = VerificationCode.objects.get_or_create(codeType='Reset', code=generated_value, userId=user)
+                send_mail(
+                    'Password Reset',
+                    f'Hello, {user.first_name}. This is your password reset code {generated_value}.\
+                        Enter this into the field provided and set a new password',
+                    'admin@admin.com',
+                    [email],
+                    fail_silently=False,
+                )
+                return Response(status=status.HTTP_202_ACCEPTED)
+            return Response(status=status.HTTP_400_BAD_REQUEST)
